requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3

cryptography~=2.8
pip~=22.3.1
keyring~=18.0.1
colorama~=0.4.3
psycopg2~=2.9.5
PyYAML~=5.3.1
Jinja2~=2.10.1
numpy~=1.23.5
requests~=2.28.1
MarkupSafe~=1.1.0
certifi~=2022.9.24
itypes~=1.2.0
coreapi~=2.3.3
coreschema~=0.0.4
click~=7.1.2
uritemplate~=4.1.1
pytest~=4.6.9
decorator~=4.4.2
matplotlib~=3.1.2
code~=0.1.0
packaging~=23.0
pexpect~=4.6.0
setuptools~=45.2.0
wcwidth~=0.1.8
PyGObject~=3.36.0
six~=1.14.0
inflection~=0.5.1
Pillow~=7.0.0
idna~=3.4
urllib3~=1.26.13
chardet~=3.0.4
simplejson~=3.16.0
PyJWT~=1.7.1
celery~=5.2.7